+++ xxl-job-core/src/main/java/com/xxl/job/core/biz/model/TriggerParam.java
@@ -8,7 +8,7 @@ import java.io.Serializable;
 public class TriggerParam implements Serializable{
     private static final long serialVersionUID = 42L;
 
-    private int jobId;
+    private Long jobId;
 
     private String executorHandler;
     private String executorParams;
@@ -26,11 +26,11 @@ public class TriggerParam implements Serializable{
     private int broadcastTotal;
 
 
-    public int getJobId() {
+    public Long getJobId() {
         return jobId;
     }
 
-    public void setJobId(int jobId) {
+    public void setJobId(Long jobId) {
         this.jobId = jobId;
     }
